refactored directories to isolate database files, made more table models, prepared for integration with Mysql database and deployment database
