+++ flask-api/requirements.txt
@@ -3,3 +3,6 @@ python-dotenv
 music21
 urllib3==1.26.6
 SQLAlchemy
+Flask-SQLAlchemy
+pymysql
+Flask-Migrate
